Add employee pagination by job title ("oficio")
Implemented paginated search for employees by job title, including new GET/POST actions in PaginacionController, repository methods, and a dedicated view. Added ModelEmpleadoOficio to encapsulate results and total count. Updated navigation to use a dropdown with the new feature. Changed Empleado.Salario type from decimal to int. Added SQL comments for the new stored procedure.

+++ MvcCorePaginacionRegistros/Models/Empleado.cs
@@ -17,7 +17,7 @@ namespace MvcCorePaginacionRegistros.Models
         public string Oficio { get; set; }
 
         [Column("SALARIO")]
-        public decimal Salario { get; set; }
+        public int Salario { get; set; }
 
         [Column("DEPT_NO")]
         public int IdDepartamento { get; set; }
